many bugs fixed
git-svn-id: http://co19.googlecode.com/svn/trunk/co19/tests/co19@561 2ecb45c1-dfb3-dc95-330a-13f216d23a04

+++ src/LibTest/core/int/ceilToDouble_A01_t01.dart
@@ -15,7 +15,7 @@ import "../../../Utils/expect.dart";
 
 import "dart:math" as Math;
 
-void check(double x) {
+void check(int x) {
   Expect.equals(x.toDouble(), x.ceilToDouble());
 }
